backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
pydantic>=2.6
email-validator>=2.1
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt>=3.2,<4.0
python-multipart>=0.0.9
